content/media/test/test_paused_after_ended.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: paused</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var manager = new MediaTestManager;

function ended(evt) {
  var v = evt.target;
  is(v.paused, false, v._name + " must not be paused after end");
  manager.finished(v.token);
}

function startTest(test, token) {
  var v = document.createElement('video');
  v.token = token;
  manager.started(v.token);
  v.src = test.name;
  v._name = test.name;
  v._finished = false;
  v.load();
  is(v.paused, true,  v._name + " must be paused at start");

  v.play();
  is(v.paused, false, v._name + " must not be paused after play");

  v.addEventListener("ended", ended, false);
}

manager.runTests(gPausedAfterEndedTests, startTest);

</script>
</pre>
</body>
</html>
